+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 All notable changes to this project will be documented in this file.
 
 ## 0.10.1
+### Added
+- Option to display attributes in _Data Model Editor_ in groups
 ### Fixed
 - Frontend errors due to wrong dayjs import
 - Missing translation for _milligram_ in si-unit attribute
+### Changed
+- Move switch to show hidden attributes in _Data Model Editor_ to dropdown
 
 ## 0.10 - Jelling
 ### Added